refactor: 메타데이터 별도 폴더에 분리, 안쓰이는 컴포넌트 삭제제
