--- app/(anon)/_components/common/stateIcon/StateIcon.tsx
@@ -1,38 +1,0 @@
-'use client';
-
-import React from 'react';
-import { styles } from './StateIcon.styles';
-import StateIconItem from './StateIconItem';
-import { useGetStepResult } from '@/hooks/useStepResultQueries';
-
-interface StateIconProps {
-  stepNumber: string;
-  userAddressNickname: string;
-  initialData?: {
-    match: number;
-    mismatch: number;
-    unchecked: number;
-  };
-}
-
-export default function StateIcon({ stepNumber, userAddressNickname, initialData }: StateIconProps) {
-  // stepResults 데이터 가져오기
-  const { data: stepData, isLoading } = useGetStepResult({
-    userAddressNickname,
-    stepNumber,
-    detail: '',
-  });
-
-  // 초기 데이터가 있으면 사용하고, 없으면 CSR 데이터 사용
-  const match = (stepData as { summary: { totalMatch: number; totalUnchecked: number; totalMismatch: number } })?.summary?.totalMatch ?? initialData?.match ?? 0;
-  const unchecked = (stepData as { summary: { totalMatch: number; totalUnchecked: number; totalMismatch: number } })?.summary?.totalUnchecked ?? initialData?.unchecked ?? 0;
-  const mismatch = (stepData as { summary: { totalMatch: number; totalUnchecked: number; totalMismatch: number } })?.summary?.totalMismatch ?? initialData?.mismatch ?? 0;
-
-  return (
-    <div className={styles.container}>
-      <StateIconItem type='match' count={match} />
-      <StateIconItem type='unchecked' count={unchecked} />
-      <StateIconItem type='mismatch' count={mismatch} />
-    </div>
-  );
-}
